Update PSCode Interpreter.cpp
I deleted the code in Parser that handled mathematical expressions and moved it into the VariableExpression class. I then generated methods in VariableExpression that would simplify the stream of tokens, create an AST, and evaluate the expression to generate a result. I also added a destructor to VariableExpression and edited the code in Parser.variableExpression to fit the requirements of the new VariableExpression class.